feat: update registration code
