+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "tidebit-defi",
-  "version": "0.9.0+89",
+  "version": "0.9.0+90",
   "private": true,
   "type": "module",
   "scripts": {
