+++ src/components/login/register_modal.tsx
@@ -8,21 +8,25 @@ import { RxCross2 } from 'react-icons/rx';
 import { useGlobalCtx } from '@/contexts/global_context';
 import { useModalCtx } from '@/contexts/modal_context';
 import { Button } from '@/components/common/button';
-import { encodeFunctionData, type Hex } from 'viem';
-import { publicClient } from '@/lib/viem';
-import { CONTRACT_ADDRESSES, ABIS } from '@/config/contracts';
-import {
-  fido2ClientService,
-  getRegisterChallenge,
-  parsePasskey,
-  sendUserOpToBundler,
-} from '@/lib/auth/fido2_client';
-import { encodeWebAuthnSignature, hexToBase64Url } from '@/lib/auth/crypto_utils';
+import { registrationService, RegistrationStep } from '@/services/registration.service';
 
 const RegisterModal: React.FC = () => {
   const [inputValue, setInputValue] = useState<string>('');
   const [isLoading, setIsLoading] = useState<boolean>(false);
-  const [stepInfo, setStepInfo] = useState<string>('Sign Up');
+  const [currentStep, setCurrentStep] = useState<RegistrationStep>('IDLE');
+
+  const stepTextMap: Record<RegistrationStep, string> = {
+    IDLE: 'Sign Up',
+    FETCHING_CHALLENGE: 'Getting secure challenge...',
+    CREATING_PASSKEY: 'Creating your biometric key...',
+    PARSING_PASSKEY: 'Parsing credentials...',
+    PREDICTING_ADDRESS: 'Predicting your wallet address...',
+    CALCULATING_HASH: 'Preparing deployment data...',
+    AWAITING_SIGNATURE: 'Please sign to authorize account creation...',
+    DEPLOYING: 'Submitting to blockchain...',
+    SUCCESS: 'Account Created!',
+    FAILED: 'Registration failed',
+  };
 
   const {
     isRegisterModalVisible: isModalVisible,
@@ -49,123 +53,20 @@ const RegisterModal: React.FC = () => {
   // Info: (20251223 - Tzuhan) 核心註冊邏輯
   const handleSignUp = async () => {
     setIsLoading(true);
-    setStepInfo('Creating Passkey...');
 
     try {
-      const username = inputValue.trim();
-      const imageUrl = 'default_avatar_url';
-
-      // Info: (20251223 - Tzuhan) --- 步驟 1: 註冊 Passkey (獲取公鑰) ---
-      // Info: (20251223 - Tzuhan) ★★★ 關鍵修正：使用瀏覽器原生 API 產生 Challenge (避開 Buffer) ★★★
-      const regChallenge = await getRegisterChallenge();
-
-      const registration = await fido2ClientService.startRegistration({
-        user: username,
-        challenge: regChallenge,
-        userVerification: 'required',
-        discoverable: 'preferred',
-      });
-
-      // Info: (20251223 - Tzuhan) 3. 呼叫後端解析，取得 P-256 公鑰座標 (X, Y)
-      const { x, y, credentialID } = await parsePasskey(registration, regChallenge);
-      console.log('Parsed Key:', { x, y, credentialID });
-
-      // Info: (20251223 - Tzuhan) 4. 準備合約部署參數
-      const salt = BigInt(0); // Info: (20251223 - Tzuhan) 這裡先用 0，實務上可用隨機數
-      const pubKeyX = BigInt(x);
-      const pubKeyY = BigInt(y);
-
-      if (!CONTRACT_ADDRESSES.FACTORY) throw new Error('Factory Address not set');
-
-      // Info: (20251223 - Tzuhan) 5. 預測未來的 SCW 地址 (呼叫 Factory 的 view function)
-      const scwAddress = await publicClient.readContract({
-        address: CONTRACT_ADDRESSES.FACTORY,
-        abi: ABIS.FACTORY,
-        functionName: 'getAddress',
-        args: [pubKeyX, pubKeyY, salt],
-      });
-      console.log('Predicted SCW Address:', scwAddress);
-
-      // Info: (20251223 - Tzuhan) 6. 組裝 UserOp 的 initCode (Factory Address + createAccount encoded data)
-      // Info: (20251226 - Tzuhan) Update: 這裡加入 username 和 imageUrl
-      const factoryCallData = encodeFunctionData({
-        abi: ABIS.FACTORY,
-        functionName: 'createAccount',
-        args: [pubKeyX, pubKeyY, salt, credentialID, username, imageUrl],
-      });
-      const initCode = `${CONTRACT_ADDRESSES.FACTORY}${factoryCallData.slice(2)}` as Hex;
-
-      const partialUserOp = {
-        sender: scwAddress,
-        nonce: BigInt(0),
-        initCode: initCode, // Info: (20251223 - Tzuhan) 部署時 nonce 通常為 0
-        callData: '0x' as Hex, // Info: (20251223 - Tzuhan) 部署時不執行其他函式
-        // Info: (20251226 - Tzuhan) Gas 設定 (部署合約需要較多 Gas)
-        callGasLimit: BigInt(200_000),
-        verificationGasLimit: BigInt(3_500_000),
-        preVerificationGas: BigInt(100_000),
-        maxFeePerGas: BigInt(0), // Info: (20251223 - Tzuhan) 0 Gas 費由 Relayer 買單
-        maxPriorityFeePerGas: BigInt(0),
-        paymasterAndData: '0x' as Hex,
-        signature: '0x' as Hex,
-      };
-
-      // Info: (20251226 - Tzuhan) --- 步驟 4: 計算 Hash ---
-      if (!CONTRACT_ADDRESSES.ENTRY_POINT) throw new Error('EntryPoint Address not set');
-
-      const userOpHash = await publicClient.readContract({
-        address: CONTRACT_ADDRESSES.ENTRY_POINT,
-        abi: ABIS.ENTRY_POINT,
-        functionName: 'getUserOpHash',
-        args: [partialUserOp],
-      });
-
-      // Info: (20251226 - Tzuhan) --- 步驟 5: 簽名 (Login 動作) ---
-      setStepInfo('Please Sign to Deploy...');
-
-      // Info: (20251226 - Tzuhan) ★★★ 關鍵：使用自定義函數將 Hex Hash 轉為 Base64URL (避開 Buffer) ★★★
-      const challengeBase64 = hexToBase64Url(userOpHash);
-
-      const authentication = await fido2ClientService.startLogin({
-        challenge: challengeBase64,
-        userVerification: 'required',
-        timeout: 60000,
-      });
-
-      // Info: (20251226 - Tzuhan) --- 步驟 6: 打包簽名並發送 ---
-      setStepInfo('Deploying...');
-      const encodedSignature = encodeWebAuthnSignature(authentication, pubKeyX, pubKeyY);
-
-      const finalUserOp = {
-        sender: partialUserOp.sender,
-        nonce: `0x${partialUserOp.nonce.toString(16)}`,
-        initCode: partialUserOp.initCode,
-        callData: partialUserOp.callData,
-        callGasLimit: `0x${partialUserOp.callGasLimit.toString(16)}`,
-        verificationGasLimit: `0x${partialUserOp.verificationGasLimit.toString(16)}`,
-        preVerificationGas: `0x${partialUserOp.preVerificationGas.toString(16)}`,
-        maxFeePerGas: `0x${partialUserOp.maxFeePerGas.toString(16)}`,
-        maxPriorityFeePerGas: `0x${partialUserOp.maxPriorityFeePerGas.toString(16)}`,
-        paymasterAndData: partialUserOp.paymasterAndData,
-        signature: encodedSignature,
-      };
-
-      const result = await sendUserOpToBundler(finalUserOp, CONTRACT_ADDRESSES.ENTRY_POINT);
-
-      if (result.code === 'SUCCESS' || result.success === true) {
-        alert(`Successfully Deployed! Address: ${scwAddress}`);
-        onClose();
-      } else {
-        throw new Error(
-          result.message || JSON.stringify(result.payload?.error) || 'Deployment failed'
-        );
-      }
+      const result = await registrationService.signUp(
+        inputValue.trim(),
+        (step) => setCurrentStep(step) // 更新 UI 狀態
+      );
+      alert(`Successfully Deployed! Address: ${result.scwAddress}`);
+      onClose();
     } catch (error) {
       console.error('Sign up failed:', error);
       alert(`Sign up failed: ${(error as Error).message}`);
     } finally {
       setIsLoading(false);
-      setStepInfo('Sign Up');
+      setCurrentStep('IDLE');
     }
   };
 
@@ -238,7 +139,7 @@ const RegisterModal: React.FC = () => {
             Cancel
           </Button>
           <Button type="button" disabled={isSubmitDisabled || isLoading} onClick={handleSignUp}>
-            {isLoading ? stepInfo : 'Sign Up'}
+            {isLoading ? stepTextMap[currentStep] : 'Sign Up'}
           </Button>
         </div>
       </div>
